refactor: redirect type import
