--- src/triggers/types.d.ts
@@ -1,4 +1,0 @@
-export type NavigationTrigger = Readonly<{
-  activate(): void;
-  inactivate(): void;
-}>;
